XSW-8607: x2: optoe: increase write timeout to 500 ms
Add setting of default optoe write timeout to 500 ms to all transceivers
for EVB and Accton platforms.

The 500 ms timeout was selected as it corresponds to the maximum timeout
specified in the optoe driver patch in the sonic-linux-kernel submodule
at the "Dynamic write timeout support for optoe driver" commit.

This is done because there are cases where the transceiver
holds off access for up to 4960[ms], according to the CMIS
standard, such as with CDB commands.

Signed-off-by: Asher Oved <ashero@xsightlabs.com>