SON-2028: xplt: fix eth0 connectivity loss after reboot
The eth0 interface intermittently disappears on reboot.
When this happens, eth0 is assigned by ice driver to different
(0000:f4:00.0) interface, which concurrently comes up first.
The udev tries to fix it later but fails to rename the interface
since it already exists. The eth0 is released later, but no one
tries to reuse it again.

sonic systemd-udevd[380]: eth1: Failed to rename network interface 3 from 'eth1' to 'eth0': File exists
sonic kernel: ice 0000:f4:00.0 eth10g0: renamed from eth0

As a workaround, we reload the igb driver when the eth0 was busy
during the igb probe.

Signed-off-by: Michael Zaidman <michaelz@xsightlabs.com>