SON-1985: es9618xx: add watchdog reboot cause indication
Detect the hardware reboot cause in the sonic_platform.chassis.get_reboot_cause(),
as most vendors do, instead of via the/host/reboot-cause/reboot-cause.txt file.

Add watchdog reboot indication to the `get_reboot_cause` routine.
The hardware reboot cause is set to the watchdog reboot cause only if the file
`/host/reboot-cause/reboot-cause.txt` is empty or contains the string `Unknown`.
This is done because the hardware watchdog reboot cause remains in the watchdog
indication during reboots that don't go through the BIOS, such as fast-reboot,
soft-reboot, and warm-reboot. Therefore, we use the above file to determine if
the reboot was caused by the watchdog or by software.

The routine `get_reboot_cause` is designed to be extendable for handling
other hardware reboot causes.

Signed-off-by: Asher Oved <ashero@xsightlabs.com>