SON-2097: es9618xx: fix fan not ok false indication
Fix false 'Not OK' status indication in the `show platform fan` command.

This issue occurs because `thermalctld` creates a completely new instance of
the `chassis.py` module, which in turn initializes its own `fan.py` module.
The `fan.py` module stores the target speed as a variable, which remains at 100%
since the target fan speed is changed in a different instance of the `fan.py` module.

The issue has been resolved by replacing the target speed variable with a temporary
file in the `/tmp` directory.

Signed-off-by: Asher Oved <ashero@xsightlabs.com>