es9632xx: fix port kernel modules to kernel 6.1.0
The same issue existed for es9618xx platform and was fixed some time
ago.

We didn't see issue with es9632xx platform package and Kernel 6.1.0
before because it was hidden due to Makefile bug.

Makefile bug was fixed and revealed this issue for es9632xx platform and
Kernel 6.1.0 also.

Makefile bug was with this commit
es9618xx: fix port kernel modules to kernel 6.1.0