XSW-10893: es9618xx: fix invalid escape sequence warning
The sequence `\s` is not a standard Python string escape sequence.
However, it is a standard escape sequence in regular expressions (regex)
to match a whitespace character.

When Python's compiler sees `\s` in a regular string, it doesn't interpret
it as a special character but issues a warning to prompt you to fix the
potential issue before it becomes an error in an upcoming Python release.

It can be fixed by using a raw string literal `r'()`, making the code
future-proof and preventing the warning.

The printed warning is:
DeprecationWarning: invalid escape sequence \s
'(0x[0-9a-fA-F]{2})([\s]+[\S]+[\s]+)([\S]+)', line)

Signed-off-by: Asher Oved <ashero@xsightlabs.com>